+++ doc/releasenotes.rst
@@ -18,7 +18,9 @@ Release Notes
 0.9.10 (2019-03-27)
 -------------------
 
-* Save number of samples in calibration step in database.
+* Save number of samples taken in calibration step in database.
+* Fix error with reported number of simulation in EpsMixin based samplers.
+* Fix several minor issues.
 
 
 0.9.9 (2019-03-25)